pyproject.toml
[tool.poetry]
name = "djancontact"
version = "0.1.0"
description = ""
authors = ["Mohammad sh <sh.mohammad66@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"
psycopg = {extras = ["binary"], version = "^3.1.19"}
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.4"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
